requirements.txt
discord.py>=2.0.0
requests>=2.25.0
python-dotenv>=0.15.0
aiohttp>=3.7.4
python-dateutil>=2.8.1 